refactor: split mongodb.go by sync stage
