+++ pkg/syncer/mongodb/checkpoint.go
@@ -1,0 +1,89 @@
+package mongodb
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"go.mongodb.org/mongo-driver/bson"
+)
+
+func (s *MongoDBSyncer) loadMongoDBResumeToken(db, coll string) bson.Raw {
+	key := fmt.Sprintf("%s.%s", db, coll)
+
+	s.resumeTokensM.RLock()
+	if token, exists := s.resumeTokens[key]; exists {
+		s.resumeTokensM.RUnlock()
+		return token
+	}
+	s.resumeTokensM.RUnlock()
+
+	if s.cfg.MongoDBResumeTokenPath == "" {
+		return nil
+	}
+
+	path := s.getResumeTokenPath(db, coll)
+	data, err := os.ReadFile(path)
+	if err != nil {
+		return nil
+	}
+	if len(data) <= 1 {
+		return nil
+	}
+
+	var token bson.Raw
+	if errU := json.Unmarshal(data, &token); errU != nil {
+		s.logger.Errorf("[MongoDB] unmarshal resume token fail => %v", errU)
+		s.removeMongoDBResumeToken(db, coll)
+		return nil
+	}
+
+	s.resumeTokensM.Lock()
+	s.resumeTokens[key] = token
+	s.resumeTokensM.Unlock()
+
+	return token
+}
+
+func (s *MongoDBSyncer) saveMongoDBResumeToken(db, coll string, token bson.Raw) {
+	if token == nil {
+		return
+	}
+
+	key := fmt.Sprintf("%s.%s", db, coll)
+	s.resumeTokensM.Lock()
+	s.resumeTokens[key] = token
+	s.resumeTokensM.Unlock()
+
+	if s.cfg.MongoDBResumeTokenPath == "" {
+		return
+	}
+
+	path := s.getResumeTokenPath(db, coll)
+	data, err := json.Marshal(token)
+	if err != nil {
+		s.logger.Errorf("[MongoDB] marshal resume token fail => %v", err)
+		return
+	}
+	if errW := os.WriteFile(path, data, 0644); errW != nil {
+		s.logger.Errorf("[MongoDB] write resume token file => %v", errW)
+	}
+}
+
+func (s *MongoDBSyncer) removeMongoDBResumeToken(db, coll string) {
+	if s.cfg.MongoDBResumeTokenPath == "" {
+		return
+	}
+	path := s.getResumeTokenPath(db, coll)
+	_ = os.Remove(path)
+	s.resumeTokensM.Lock()
+	delete(s.resumeTokens, fmt.Sprintf("%s.%s", db, coll))
+	s.resumeTokensM.Unlock()
+	s.logger.Infof("[MongoDB] removed invalid resume token => %s", path)
+}
+
+func (s *MongoDBSyncer) getResumeTokenPath(db, coll string) string {
+	fileName := fmt.Sprintf("%s_%s.json", db, coll)
+	return filepath.Join(s.cfg.MongoDBResumeTokenPath, fileName)
+}
