+++ pkg/syncer/mongodb/snapshot.go
@@ -1,0 +1,79 @@
+package mongodb
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/retail-ai-inc/sync/pkg/utils"
+	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/mongo"
+)
+
+func (s *MongoDBSyncer) doInitialSync(ctx context.Context, sourceColl, targetColl *mongo.Collection, sourceDB, targetDB string) error {
+	count, err := targetColl.EstimatedDocumentCount(ctx)
+	if err != nil {
+		return fmt.Errorf("check target collection %s.%s fail: %v", targetDB, targetColl.Name(), err)
+	}
+	if count > 0 {
+		s.logger.Infof("[MongoDB] %s.%s has data => skip initial sync", targetDB, targetColl.Name())
+		return nil
+	}
+
+	s.logger.Infof("[MongoDB] Starting initial sync for %s.%s -> %s.%s", sourceDB, sourceColl.Name(), targetDB, targetColl.Name())
+
+	cursor, err := sourceColl.Find(ctx, bson.M{})
+	if err != nil {
+		return fmt.Errorf("source find fail => %v", err)
+	}
+	defer cursor.Close(ctx)
+
+	batchSize := 100
+	var batch []interface{}
+	inserted := 0
+
+	for cursor.Next(ctx) {
+		var doc bson.M
+		if errD := cursor.Decode(&doc); errD != nil {
+			return fmt.Errorf("decode doc fail => %v", errD)
+		}
+		batch = append(batch, doc)
+		if len(batch) >= batchSize {
+			err := utils.RetryMongoOperation(ctx, s.logger,
+				fmt.Sprintf("InsertMany to %s.%s", targetDB, targetColl.Name()),
+				func() error {
+					res, err := targetColl.InsertMany(ctx, batch)
+					if err != nil {
+						return err
+					}
+					inserted += len(res.InsertedIDs)
+					return nil
+				})
+
+			if err != nil {
+				return fmt.Errorf("insertMany fail => %v", err)
+			}
+			batch = batch[:0]
+		}
+	}
+
+	if len(batch) > 0 {
+		err := utils.RetryMongoOperation(ctx, s.logger,
+			fmt.Sprintf("InsertMany final batch to %s.%s", targetDB, targetColl.Name()),
+			func() error {
+				res, err := targetColl.InsertMany(ctx, batch)
+				if err != nil {
+					return err
+				}
+				inserted += len(res.InsertedIDs)
+				return nil
+			})
+
+		if err != nil {
+			return fmt.Errorf("insertMany fail => %v", err)
+		}
+	}
+
+	s.logger.Infof("[MongoDB] doInitialSync => %s.%s => %s.%s inserted=%d docs",
+		sourceDB, sourceColl.Name(), targetDB, targetColl.Name(), inserted)
+	return nil
+}
